requirements.txt
pycryptodomex==3.15.0
PyExecJS==1.5.1
requests==2.28.1
